refactor: extract ipc protocol helpers
