+++ internal/il2cpp-dll-injection.vcxproj.filters
@@ -11,6 +11,10 @@
     <ClCompile Include="src\bootstrap\version.cpp" />
     <ClCompile Include="src\platform\dx11\Dx11.cpp" />
     <ClCompile Include="src\core\ipc\IpcBridge.cpp" />
+    <ClCompile Include="src\core\ipc\IpcFraming.cpp" />
+    <ClCompile Include="src\core\ipc\IpcJson.cpp" />
+    <ClCompile Include="src\core\ipc\IpcMessages.cpp" />
+    <ClCompile Include="src\core\ipc\IpcSession.cpp" />
     <ClCompile Include="src\core\ipc\IpcTileState.cpp" />
     <ClCompile Include="src\core\ipc\Handshake.cpp" />
     <ClCompile Include="src\ui\gui\tabs\CameraTAB.cpp" />
@@ -84,6 +88,10 @@
     <ClInclude Include="src\bootstrap\version.h" />
     <ClInclude Include="src\platform\dx11\Dx11.h" />
     <ClInclude Include="src\core\ipc\IpcBridge.h" />
+    <ClInclude Include="src\core\ipc\IpcFraming.h" />
+    <ClInclude Include="src\core\ipc\IpcJson.h" />
+    <ClInclude Include="src\core\ipc\IpcMessages.h" />
+    <ClInclude Include="src\core\ipc\IpcSession.h" />
     <ClInclude Include="src\core\ipc\IpcTileState.h" />
     <ClInclude Include="src\core\ipc\IpcBridgeIncludes.hpp" />
     <ClInclude Include="src\ui\IpcBridge.h" />
