+++ internal/src/core/ipc/IpcFraming.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <Windows.h>
+
+namespace IpcFraming {
+
+bool WriteMessage(HANDLE hPipe, const char* json, int len);
+int ReadMessage(HANDLE hPipe, char* buf, int bufSize);
+
+} // namespace IpcFraming
