+++ internal/src/core/ipc/IpcJson.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace IpcJson {
+
+char* GetString(char* json, const char* key, char* valBuf, int valBufSize);
+bool GetBool(char* json, const char* key);
+bool GetNumberToken(char* json, const char* key, char* outBuf, int outBufSize);
+
+} // namespace IpcJson
